refactor: 💡 pull the team names from rbac
