+++ routes/mutate.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
+	n "github.com/ministryofjustice/cloud-platform-label-pods/pkg/get_team"
 	m "github.com/ministryofjustice/cloud-platform-label-pods/pkg/mutate"
-	n "github.com/ministryofjustice/cloud-platform-label-pods/pkg/namespace"
 	"github.com/ministryofjustice/cloud-platform-label-pods/utils"
 )
 
@@ -24,7 +24,7 @@ func initMutatePod(r *gin.Engine) {
 			utils.SendResponse(c, errObj)
 		}
 
-		getGithubTeamnameFn := n.InitGetGithubTeamName(n.GetTeamNameFromNs)
+		getGithubTeamnameFn := n.InitGetGithubTeamName(n.GetTeamName)
 
 		mutated, err := m.Mutate(body, getGithubTeamnameFn)
 		if err != nil {
